The emscripten commit, read futher:
Removes SDL threads because it was whining on the wasm build (I probaly just had to compile SDL with the -pthread flag but that would be quite a headache)

WASM performance is really disapointing, only a 640x360 h.264 clip runned smothly even with -O3 optimazations. So I decided to stop trying to make this run in the web, but the wasm preset and configurations are stil avaiable if I change my mind. I guess again that the performance could be improved by having the right compiler flags

If you want to compile this to wasm, make sure to change the input file name in the Player.hpp and in preload file in the src/CMakeLists.txt. This file must be in the build/vcpkg-emscripten/src folder

also added the CMakePresets.json file

Even though not much has changed, I've spent around 16 hours trying to make this compile, so yeah, cool